backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-docx>=1.1
docxcompose>=1.4
pydantic>=2.0
pytest>=8.0
